requirements.txt
django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
redis==5.2.1
bcrypt==4.3.0
django-cors-headers==4.7.0
django-filter==25.1
drf-yasg==1.21.10
pillow==11.2.1
pytz==2025.2
django-axes==7.0.2
python-decouple==3.8 